restructured components in folders
