--- fitnessappfrontend/src/components/empty.component.tsx
@@ -1,7 +1,0 @@
-import React, { Component } from 'react';
-
-export class EmptyComponent extends React.Component<any, any> {
-  render() {
-    return undefined;
-  }
-}
