+++ fitnessappfrontend/src/components/logout/logout.component.tsx
