+++ fitnessappfrontend/src/components/nav/nav.component.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-const logo = require('../resources/fitness-icon.png');
+const logo = require('../../resources/fitness-icon.png');
 // import logo from '../resources/fitness-pattern-blue.jpg' // relative path to image
 import { AxiosResponse } from 'axios';
 import { Link } from 'react-router-dom';
